add filereader
