+++ CoreWinSubLog/LogRecord.cs
@@ -82,41 +82,41 @@ namespace CoreWinSubLog
         //    ModuleName = moduleName;
         //}
 
-        //Commit by Yuqing as the string doesn't contain an integral information of log record.
-        /// <summary>
-        /// Get a log record form given string
-        /// </summary>
-        /// <param name="content">string matched to a log record</param>
-        /// <returns></returns>
-        public static LogRecord FromString(string content)
-        {
-            if (string.IsNullOrEmpty(content))
-            {
-                return NullRecord();
-            }
+        ////Commit by Yuqing as the string doesn't contain an integral information of log record.
+        ///// <summary>
+        ///// Get a log record form given string
+        ///// </summary>
+        ///// <param name="content">string matched to a log record</param>
+        ///// <returns></returns>
+        //public static LogRecord FromString(string content)
+        //{
+        //    if (string.IsNullOrEmpty(content))
+        //    {
+        //        return NullRecord();
+        //    }
 
-            string timeString = content.Substring(0, 23);
-            DateTime dateTime;
-            DateTime.TryParseExact(timeString, "yyyy-MM-dd HH:mm:ss.fff", CultureInfo.CurrentCulture, DateTimeStyles.None, out dateTime);
+        //    string timeString = content.Substring(0, 23);
+        //    DateTime dateTime;
+        //    DateTime.TryParseExact(timeString, "yyyy-MM-dd HH:mm:ss.fff", CultureInfo.CurrentCulture, DateTimeStyles.None, out dateTime);
 
-            string otherString = content.Substring(24);
+        //    string otherString = content.Substring(24);
 
-            LogLevel logLevel = LogLevel.Info;
-            string theMessage = string.Empty;
-            if (otherString.StartsWith("["))
-            {
-                string[] msgs = otherString.Split(new string[] { " " }, 2, StringSplitOptions.None);
-                string levelString = msgs[0].TrimStart('[').TrimEnd(']');
-                Enum.TryParse(levelString, out logLevel);
-                theMessage = msgs[1];
-            }
-            else
-            {
-                theMessage = otherString;
-            }
+        //    LogLevel logLevel = LogLevel.Info;
+        //    string theMessage = string.Empty;
+        //    if (otherString.StartsWith("["))
+        //    {
+        //        string[] msgs = otherString.Split(new string[] { " " }, 2, StringSplitOptions.None);
+        //        string levelString = msgs[0].TrimStart('[').TrimEnd(']');
+        //        Enum.TryParse(levelString, out logLevel);
+        //        theMessage = msgs[1];
+        //    }
+        //    else
+        //    {
+        //        theMessage = otherString;
+        //    }
 
-            return new LogRecord(logLevel, dateTime, "", theMessage);
-        }
+        //    return new LogRecord(logLevel, dateTime, "", theMessage);
+        //}
 
         /// <summary>
         /// Convert log record to string (DateTime ("yyyy-MM-dd HH:mm:ss.fff")).
