+++ LoggingSystemTest/Program.cs
@@ -16,14 +16,15 @@ namespace LoggingSystemTest
         {
             try
             {
-                LoggerTest.TestNullLogger();
-                Console.WriteLine();
-                LoggerTest.TestConsoleLogger();
-                LoggerTest.TestWcfLogger();
-                //LoggerTest.TestLogRecordString();
-                LoggerTest.TestWriteFile();
-                LoggerTest.TestAsyWriteFile();
-                LoggerTest.TestDeleteFirstLine();
+                //LoggerTest.TestNullLogger();
+                //Console.WriteLine();
+                //LoggerTest.TestConsoleLogger();
+                //LoggerTest.TestWcfLogger();
+                ////LoggerTest.TestLogRecordString();
+                //LoggerTest.TestWriteFile();
+                //LoggerTest.TestAsyWriteFile();
+                //LoggerTest.TestDeleteFirstLine();
+                LoggerTest.TestReadLine();
             }
             catch (Exception ex)
             {
